refactor: co-locate charts components in charts subfolder
